refactor(typechecker): extract pathSegments.ts to break flow<->narrowing cycle
M2 added a value import of chainToSegments from flow.ts into narrowing.ts, while
flow.ts value-imports narrowByRefine from narrowing.ts — a runtime circular
dependency. Move the pure path-segment core (PathSegment/Reference types,
segKey/referenceKey/isPrefixOf, toSegment/chainToSegments/stablePrefix) into a
leaf module pathSegments.ts that both import from; flow.ts re-exports it so
existing importers are unaffected. narrowing.ts no longer imports flow.ts.
